feature (PerksList) created Input and Row components; added token and quantity input fields; added Perk type definitions
